Move card components to proper local folders and remove stupid shit
