--- app/src/main/java/com/ilikeincest/food4student/AppViewModel.kt
@@ -1,38 +1,0 @@
-package com.ilikeincest.food4student
-
-import android.util.Log
-import androidx.credentials.Credential
-import androidx.credentials.CustomCredential
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import com.google.android.libraries.identity.googleid.GoogleIdTokenCredential
-import com.google.android.libraries.identity.googleid.GoogleIdTokenCredential.Companion.TYPE_GOOGLE_ID_TOKEN_CREDENTIAL
-import com.ilikeincest.food4student.model.service.AccountService
-import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.CoroutineExceptionHandler
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.launch
-import javax.inject.Inject
-
-@HiltViewModel
-class AppViewModel @Inject constructor(
-    private val accountService: AccountService
-) : ViewModel() {
-    fun launchCatching(block: suspend CoroutineScope.() -> Unit) =
-        viewModelScope.launch(
-            CoroutineExceptionHandler { _, throwable ->
-                Log.d(ACCOUNT_ERROR_TAG, throwable.message.orEmpty())
-            },
-            block = block
-        )
-    fun onSignInWithGoogle(credential: Credential) {
-        launchCatching {
-            if (credential is CustomCredential && credential.type == TYPE_GOOGLE_ID_TOKEN_CREDENTIAL) {
-                val googleIdTokenCredential = GoogleIdTokenCredential.createFrom(credential.data)
-                accountService.signInWithGoogle(googleIdTokenCredential.idToken)
-            } else {
-                Log.e(ACCOUNT_ERROR_TAG, "Unexpected credentials")
-            }
-        }
-    }
-}
